Explicitly try other EAs to guarantee the correct EA is reached
Partially resolves issue #1.
Convergence is still not guaranteed, as if an orbital is unbound,
it's still impossble to just force it to be occupied. If it does
converge however, I can be quite confident that it converges to the
correct value. It's very thorough in the way it checks nearby
arrangements, and could be made much faster with some heuristics.