Avoid rounding errors
Scale the inner edge of the orbital using atomic number to compensate
for the change in potential, so that all orbitals have increased by a
similar amount by the point where the nuclear potential becomes
negative. This reduces the orbital's dependence on the energy, preventing
the limited precision from preventing the energy converging properly.