Improve exchange energy, and minor tweaks
Resolve #4 and #5. Exchange energy in atomEnergy is now perfect except
that each orbital is an exact L^2, Lxy & Lzw eigenstate and it's
averaged within each sub-shell, and the exchange energy in Atom.hs
includes only the leading-order term, which should only be a very
minor error when it's at least p-p, in which case exchange energy
is very small relative to the coulumb term anyway. Functions were
added to calculate these coefficients, but as they turned out to have
a simple form, that was used directly instead.