Allow loss of electrons
By allowing alternative EAs with fewer electrons but better energies
to be considered, it's more likely to find the correct EA, and in some
cases find it more quickly. This probably resolves #11, but it's hard
to tell as there seem to be other things interfering with that. This
also makes more physical sense, as if a state is higher in energy than
one with fewer electrons, it's not really a bound ground state
(equivalently, in the reference state (gas phase 0K), the electron
chemical potential should be 0).