Refactor Atom.hs
Replace nested lists with a PerOrbital ADT for data that's stored for
each orbital in an atom. Remove prevOccs. Split Atom type in two, in
preparation for #10. Re-order functions to be more logical. Remove the
possibility of an atom's electron arrangement being determined by its
energies, completing the replacement of this with relaxAtom, thus
facilitating lots of simplification. Small changes to other files to
accommodate these changes.