Improve orbital generation accuracy
Improve the iteration used to solve the differential equation for
each orbital to one with quadratic error. Use explicit error estimates
to determine the best inner radius to start at, so that the combination
of error at the nucleus and iteration error is about optimal given the
amount of computation time used.